requirements.txt
streamlit
pillow
numpy
pyttsx3
gTTS
